% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax-package.R
\docType{package}
\name{arktax-package}
\alias{arktax}
\alias{arktax-package}
\title{arktax: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/arktax}
  \item Report bugs at \url{https://github.com/andrewallenbruce/arktax/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
