% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_footnotes}
\alias{crosswalk_footnotes}
\title{Medicare Specialty Crosswalk Footnotes}
\usage{
crosswalk_footnotes(taxonomy_code = NULL, specialty_code = NULL)
}
\arguments{
\item{taxonomy_code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}

\item{specialty_code}{\verb{<chr>}  Medicare Specialty Code, an alphanumeric code,
two characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Medicare Specialty Crosswalk Footnotes
}
\examples{
crosswalk_footnotes(taxonomy_code = "251E00000X")

crosswalk_footnotes(specialty_code = "A0")

}
