% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_raw}
\alias{crosswalk_raw}
\title{Raw Crosswalk File}
\usage{
crosswalk_raw(code = NULL)
}
\arguments{
\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Raw Crosswalk File
}
\examples{
crosswalk_raw(code = "103T00000X")

head(crosswalk_raw(), 10)

}
