% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_taxonomy}
\alias{crosswalk_taxonomy}
\title{Taxonomy - Medicare Specialty Crosswalk}
\usage{
crosswalk_taxonomy(taxonomy_code = NULL, specialty_code = NULL)
}
\arguments{
\item{taxonomy_code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}

\item{specialty_code}{\verb{<chr>}  Medicare Specialty Code, an alphanumeric code,
two characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Taxonomy - Medicare Specialty Crosswalk
}
\examples{
crosswalk_taxonomy(taxonomy_code = "103T00000X")

crosswalk_taxonomy(taxonomy_code = c("101YM0800X", "101YP2500X"))

}
