% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provider_types}
\alias{provider_types}
\title{Provider Types Reference}
\format{
\subsection{\code{provider_types}}{

A data frame with 305 rows and 5 columns:
\describe{
\item{code}{Provider Type Code (Full)}
\item{type}{Provider Type Code (Type)}
\item{type_description}{Type Description}
\item{spec}{Provider Type Code (Specialty)}
\item{spec_description}{Specialty Description}
...
}
}
}
\source{
\href{https://data.cms.gov/sites/default/files/2025-10/1e42b271-dfe0-4c64-8022-0c507993667e/PPEF_Data_Dictionary.pdf}{CMS: PPEF Data Dictionary}
}
\usage{
provider_types
}
\description{
Provider Types Reference
}
\keyword{internal}
