% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_ark}
\alias{retrieve_ark}
\title{Get Taxomony Source File}
\usage{
retrieve_ark(year = 2024, code = NULL)
}
\arguments{
\item{year}{\verb{<int>} year of rvu source file; default is \code{2020}}

\item{code}{\verb{<chr>} Taxonomy code}
}
\value{
\verb{<tibble>} of search results
}
\description{
Get Taxomony Source File
}
\examples{
retrieve_ark(2024, "101Y00000X")

}
