% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_changelog}
\alias{retrieve_changelog}
\title{Retrieve Taxonomy Changelog}
\usage{
retrieve_changelog(code = NULL)
}
\arguments{
\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Retrieve Taxonomy Changelog
}
\examples{
retrieve_changelog(code = "103GC0700X")

retrieve_changelog(code = "103G00000X")

}
