% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_hierarchy}
\alias{retrieve_hierarchy}
\title{Retrieve Taxonomy Hierarchy}
\usage{
retrieve_hierarchy(code = NULL)
}
\arguments{
\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Retrieve Taxonomy Hierarchy
}
\examples{
retrieve_hierarchy(code = "101Y00000X")

retrieve_hierarchy(code = "103TA0400X")

}
