% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_raw}
\alias{retrieve_raw}
\title{Retrieve Taxonomy Source File}
\usage{
retrieve_raw(year = NULL, version = NULL, code = NULL)
}
\arguments{
\item{year}{\verb{<int>} year of source file release; options are \code{2009:2024}}

\item{version}{\verb{<int>} version of source file; options are \code{0} or \code{1}}

\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Retrieve Taxonomy Source File
}
\examples{
retrieve_raw(year = 2024, code = "101Y00000X")

retrieve_raw(code = "101Y00000X")

}
