% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_sources}
\alias{retrieve_sources}
\title{Retrieve Taxonomy Sources}
\usage{
retrieve_sources(code = NULL)
}
\arguments{
\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Retrieve Taxonomy Sources
}
\examples{
retrieve_sources(code = "101Y00000X")

retrieve_sources(code = "103TA0400X")

}
