% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_changelog}
\alias{taxonomy_changelog}
\title{Taxonomy Change Log}
\usage{
taxonomy_changelog(taxonomy_code = NULL)
}
\arguments{
\item{taxonomy_code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Taxonomy Change Log
}
\examples{
taxonomy_changelog(taxonomy_code = "103GC0700X")

taxonomy_changelog(taxonomy_code = "103G00000X")

}
