% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{taxonomy_sources}
\alias{taxonomy_sources}
\title{Taxonomy Sources}
\usage{
taxonomy_sources(code = NULL)
}
\arguments{
\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Taxonomy Sources
}
\examples{
taxonomy_sources(code = "101Y00000X")

taxonomy_sources(code = "103TA0400X")

}
