% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax-package.R
\docType{package}
\name{arktax-package}
\alias{arktax}
\alias{arktax-package}
\title{arktax: NUCC Healthcare Provider Taxonomy Codes}
\description{
Programmatic interface to the NUCC Health Care Provider Taxonomy Code Set, including the raw codesets from 2009 to present, various different forms of the data, and the Medicare specialty-to-taxonomy crosswalk dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/arktax}
  \item \url{https://andrewallenbruce.github.io/arktax/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/arktax/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
