% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_ark}
\alias{retrieve_ark}
\title{Get Taxomony Source File}
\usage{
retrieve_ark(year = NULL, code = NULL, which = c("wide", "long"))
}
\arguments{
\item{year}{\verb{<int>} year of source file release; options are \code{2009:2024}}

\item{code}{\verb{<chr>} Taxonomy code}

\item{which}{\verb{<chr>} wide or long version of the taxonomy; options are \code{wide} or \code{long}}
}
\value{
\verb{<tibble>} of search results
}
\description{
Get Taxomony Source File
}
\examples{
retrieve_ark(year = 2024, code = "101Y00000X", which = "wide")

retrieve_ark(year = 2024, code = "101Y00000X", which = "long")

}
