% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{retrieve_ark}
\alias{retrieve_ark}
\title{Get Taxomony Source File}
\usage{
retrieve_ark(
  year = NULL,
  version = NULL,
  code = NULL,
  which = c("wide", "long")
)
}
\arguments{
\item{year}{\verb{<int>} year of source file release; options are \code{2009:2024}}

\item{version}{\verb{<int>} version of source file; options are \code{0} or \code{1}}

\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique alphanumeric code, ten characters in length}

\item{which}{\verb{<chr>} \code{wide} or \code{long} version of the taxonomy}
}
\value{
\verb{<tibble>} of search results
}
\description{
Get Taxomony Source File
}
\examples{
retrieve_ark(year = 2024, code = "101Y00000X", which = "wide")

retrieve_ark(code = "101Y00000X", which = "long")

}
