% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arktax.R
\name{taxonomy_definition}
\alias{taxonomy_definition}
\title{Taxonomy Definitions}
\usage{
taxonomy_definition(code = NULL)
}
\arguments{
\item{code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Taxonomy Definitions
}
\examples{
taxonomy_definition(code = "101Y00000X")

taxonomy_definition(code = "103TA0400X")

}
