% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_hierarchy}
\alias{taxonomy_hierarchy}
\title{Taxonomy Hierarchy}
\usage{
taxonomy_hierarchy(taxonomy_code = NULL)
}
\arguments{
\item{taxonomy_code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}
}
\value{
\verb{<tibble>} of search results
}
\description{
Taxonomy Hierarchy
}
\examples{
taxonomy_hierarchy(taxonomy_code = "101Y00000X")

taxonomy_hierarchy(taxonomy_code = "103TA0400X")

}
