% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_hierarchy}
\alias{taxonomy_hierarchy}
\title{Taxonomy Hierarchy}
\usage{
taxonomy_hierarchy(
  taxonomy_code = NULL,
  taxonomy_level = NULL,
  taxonomy_level_title = NULL
)
}
\arguments{
\item{taxonomy_code}{\verb{<chr>}  Health Care Provider Taxonomy code, a unique
alphanumeric code, ten characters in length}

\item{taxonomy_level}{\verb{<chr>}  Taxonomy level; options are \code{"I. Section"},
\code{"II. Grouping"}, \code{"III. Classification"} and \code{"IV. Specialization"}}

\item{taxonomy_level_title}{\verb{<chr>}  Taxonomy level title}
}
\value{
\verb{<tibble>} of search results
}
\description{
Taxonomy Hierarchy
}
\examples{
taxonomy_hierarchy(taxonomy_code = "101Y00000X")

taxonomy_hierarchy(taxonomy_code = "103TA0400X")

taxonomy_hierarchy(taxonomy_level = "I. Section")

taxonomy_hierarchy(taxonomy_level_title = "Allopathic & Osteopathic Physicians")

}
